refactor: Answer, Amount, BetAmount package 변경
