+++ src/main/java/domain/participant/Answer.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.participant;
 
 import constants.ErrorCode;
 import exception.InvalidCommandException;
