+++ src/test/java/domain/AnswerTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.participant.Answer;
 import exception.InvalidCommandException;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
